requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
groq>=0.4.0
websockets>=12.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
edge-tts>=6.1.9
aiofiles>=23.0.0
pydub>=0.25.1
opencv-python>=4.8.0.0
pytesseract>=0.3.10
yt-dlp>=2024.1.0
python-docx>=1.1.0
google-generativeai>=0.3.0
deep-translator>=1.11.4
python-bidi>=0.4.2
arabic-reshaper>=3.0.0
certifi>=2023.0.0


